Python Client cleanup part 2
